+++ example/lib/main.dart
@@ -5,104 +5,7 @@ void main() {
   runApp(const MyApp());
 }
 
-abstract class MyChannelSignal extends ChannelSignal {}
-
-class MyChannel extends StateChannel<MyChannelSignal> {
-  MyChannel() {
-    _counterState = CounterState(() => add(CounterStateSignal()));
-    _notificationState = NotificationState(() => add(NotificationStateSignal()));
-  }
-
-//signal: CounterStateSignal
-  late CounterState _counterState;
-  CounterState get counterState => _counterState;
-
-//signal: NotificationStateSignal
-  late NotificationState _notificationState;
-  NotificationState get notificationState => _notificationState;
-
-  @override
-  void initState() {
-    _counterState.initState();
-    _notificationState.initState();
-  }
-
-  @override
-  afterInitState() {
-    _counterState.afterInitState();
-    _notificationState.afterInitState();
-  }
-
-  @override
-  Future<void> dispose() async {
-    _counterState.dispose();
-    _notificationState.dispose();
-    super.dispose();
-  }
-}
-
-class CounterStateSignal extends MyChannelSignal {}
-
-class CounterState extends BaseState {
-  CounterState(void Function() onStateChanged) : super(onStateChanged);
-
-  int _count = 0;
-  int get count => _count;
-
-  @override
-  void initState() {
-    wait(signal: false);
-    _count = 0;
-  }
-
-  @override
-  Future<void> afterInitState() async {
-    await incrementFuture();
-  }
-
-  @override
-  void dispose() {}
-
-  void increment() {
-    _count = _count + 1;
-    doneSuccess();
-  }
-
-  void decrement() {
-    _count = _count - 1;
-    doneSuccess();
-  }
-
-  Future<void> incrementFuture() async {
-    try {
-      wait();
-      await Future<void>.delayed(const Duration(seconds: 5));
-      _count = _count + 1;
-
-      doneSuccess();
-    } catch (e) {
-      doneError(e.toString());
-    }
-  }
-
-  Future<void> decrementFuture() async {
-    try {
-      wait();
-      await Future<void>.delayed(const Duration(seconds: 1));
-      _count = _count - 1;
-
-      doneSuccess();
-    } catch (e) {
-      doneError(e.toString());
-    }
-  }
-}
-
-class NotificationStateSignal extends MyChannelSignal {}
-
-class NotificationState extends BaseState {
-  NotificationState(void Function() onStateChanged) : super(onStateChanged);
-
+class MySignal extends Signal {
   bool _isOpen = false;
   bool get isOpen => _isOpen;
 
@@ -116,7 +19,9 @@ class NotificationState extends BaseState {
   Future<void> afterInitState() async => await changeFuture();
 
   @override
-  dispose() {}
+  Future<void> dispose() async {
+    await super.dispose();
+  }
 
   change() {
     _isOpen = !_isOpen;
@@ -141,8 +46,8 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ChannelProvider<MyChannel>(
-        channel: (context) => MyChannel(),
+    return SignalProvider<MySignal>(
+        signal: (context) => MySignal(),
         child: const MaterialApp(
           debugShowCheckedModeBanner: false,
           home: HomePage(),
@@ -162,66 +67,24 @@ class HomePage extends StatelessWidget {
       body: ListView(
         children: <Widget>[
           const SizedBox(height: 50),
-          SizedBox(
-            height: 45,
-            width: 45,
-            child: ChannelBuilder<MyChannel, MyChannelSignal>(
-              condition: (channel, signal) => signal is CounterStateSignal,
-              builder: (context, channel, _) {
-                final state = channel.counterState;
-
-                return Stack(
-                  alignment: Alignment.topCenter,
-                  children: <Widget>[
-                    !state.success
-                        ? Text(state.error)
-                        : Text(state.count.toString(), style: const TextStyle(fontSize: 25)),
-                    if (state.busy) const CircularProgressIndicator(),
-                  ],
-                );
-              },
-            ),
-          ),
-          ChannelBuilder<MyChannel, MyChannelSignal>(
-              condition: (channel, signal) => signal is NotificationStateSignal,
-              builder: (context, channel, _) {
-                final state = channel.notificationState;
-
-                return Stack(
-                  alignment: Alignment.topCenter,
-                  children: <Widget>[
-                    TextButton(
-                      onPressed: state.busy ? null : () => state.change(),
-                      child: Text(state.isOpen ? 'Notification: On' : 'Notification: Off',
-                          style: TextStyle(fontSize: 25, color: state.isOpen ? Colors.green : Colors.red)),
-                    ),
-                    if (state.busy) const CircularProgressIndicator(),
-                  ],
-                );
-              }),
+          SignalBuilder<MySignal>(builder: (context, signal, _) {
+            return Stack(
+              alignment: Alignment.topCenter,
+              children: <Widget>[
+                TextButton(
+                  onPressed: signal.busy ? null : () => signal.change(),
+                  child: Text(signal.isOpen ? 'Notification: On' : 'Notification: Off',
+                      style: TextStyle(fontSize: 25, color: signal.isOpen ? Colors.green : Colors.red)),
+                ),
+                if (signal.busy) const CircularProgressIndicator(),
+              ],
+            );
+          }),
           const SizedBox(height: 50),
-          ChannelBuilder<MyChannel, MyChannelSignal>(
-            condition: (channel, signal) => signal is CounterStateSignal || signal is NotificationStateSignal,
-            builder: (context, channel, _) {
-              final state = channel.counterState;
-              return ElevatedButton(
-                onPressed: channel.counterState.busy || channel.notificationState.busy ? null : () => state.increment(),
-                child: const Text('CounterState: increment'),
-              );
-            },
-          ),
           ElevatedButton(
-            onPressed: () => ChannelProvider.of<MyChannel>(context).counterState.decrementFuture(),
-            child: const Text('CounterState: decrementFuture'),
-          ),
-          ElevatedButton(
-            onPressed: () => ChannelProvider.of<MyChannel>(context).notificationState.change(),
-            child: const Text('NotificationState: change'),
-          ),
-          ElevatedButton(
-            onPressed: () => ChannelProvider.of<MyChannel>(context).notificationState.changeFuture(),
-            child: const Text('NotificationState: changeFuture'),
-          ),
+            onPressed: () => SignalProvider.of<MySignal>(context).change(),
+            child: const Text('NotificationState: changed'),
+          )
         ],
       ),
     );