+++ lib/src/channelsignal.dart
@@ -1,7 +1,0 @@
-///Base class for a StateChannelSignal
-///
-/// This class must be inherited to create your own StateChannelSignal.
-///```dart
-/// abstract class MyChannelSignal extends ChannelSignal{}
-///```
-abstract class ChannelSignal {}